Bug fix
Had to switch from ints to long ints for the vector storing the ids of
wrestlers. This was causing problems on some computers before when attempting
to do comparisons due to the different interpretations of ints.

+++ include/Logic.h
@@ -11,6 +11,6 @@ void getInputSetSpd(Wrestler& sumo);
 
 void setAISpd(Wrestler& ai_sumo, Wrestler human_sumo);
 
-bool calcCollision(std::vector<int> ids, std::vector<Wrestler>& wrestlers, bool have_collided);
+bool calcCollision(std::vector<long int> ids, std::vector<Wrestler>& wrestlers, bool have_collided);
 
 #endif // LOGIC_H
